Single-run sparsification for MPI (#79)
* Refactor: Initial cleanup for MPI sparsify

- Remove commented out code
- Remove redundant deallocates
- Conform order of attributes
- Fix indentation (3 spaces)
- sparseMethod -> sparse_method

* Move task manager into sparsify

* Add exclude_duplicates

- extract duplicate exclusion into a subroutine to simplify code
- remove unused variables

* Add pointers to x, x_size, config_type

* Add sparsification with MPI

For simplicity, x and config_type are collected on root, which then runs
through the usual algorithms and finally distributes x_sparse,
config_type and associated data.

Method index_file is unsupported as it would complicate the code
immensely (data has already been reordered at read-in).

Selection of sparse points is not strictly identical between calls with
different MPI configurations, cf. changing the order of frames in XYZ.

* Catch methods CLUSTER, COVARIANCE, CUR_COVARIANCE

These methods call routines directly with the gpCoordinates custom type
making it more complicated to use the collected arrays for MPI. Support
may be added later but it would probably involve changing the signatures
of the routines to use the components instead of the full object.

* Deallocate %config_type after sparsification

Config_type are also used to exclude descriptor vectors that are
similar. They should not be used after sparsification so we deallocate
and do not distribute them back.

If they are needed after all, they should be distributed with scatterv.
This is difficult to replicate with sparse_method=file.

* Fix: allocate non-root gatherv vars

* Add cutoff_ptr

* Deallocate %sparseX_index after sparsification

As with %config_type, this does not mean much in a distributed scenario
with reordered input and is not used for sparse_method=file.

* Add covdiag_x_x_ptr

* Cleanup: Reduce redundancy with pointers

* Cleanup: use reallocate, swap order

* Cleanup: update comment

* Doc: Update MPI run

* Fix: Nullify pointers

* Catch method NONE

When tasks are distributed, the smaller n_sparseX is reserved for
the shared task of Kmm. An additional update for all config_types
depending on sparse_method would be necessary before distribution.